+++ internal/service/lzw/lzw.go
@@ -3,8 +3,8 @@ package lzw
 import "github.com/phaseant/lzw-cli/internal/dictionary"
 
 type Service interface {
-	Encode(text []byte, dict *dictionary.Dictionary) []int
-	Decode(text []int, dict *dictionary.Dictionary) []byte
+	Encode(text []byte, dict *dictionary.Dictionary) []uint64
+	Decode(text []uint64, dict *dictionary.Dictionary) []byte
 }
 
 type service struct {
